+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changelog
 - added `serialize`, `deserialize` methods to `weka.classifiers.Classifier` to simplify loading/saving model
 - added `serialize`, `deserialize` methods to `weka.clusterers.Clusterer` to simplify loading/saving model
 - added `serialize`, `deserialize` methods to `weka.filters.Filter` to simplify loading/saving filter
+- added methods `plot_rocs` and `plot_prcs` to `weka.plot.classifiers` module to plot ROC/PRC curve on same dataset
+  for multiple classifiers
 - ...